[Ionut] - Texture mipmaps are now compute before bind if they need a refresh (to avoid recalculation if they're not used) - More work on the editor: cleanup, load/save functionality, etc - Input scaling: Mouse position is now relative to the scene preview position in editor mode. (the scene preview rect is used to re-map X/Y absolute and relative mouse positions)
git-svn-id: https://xp-dev.com/svn/Divide-Framework/trunk@976 e89c6373-eeba-4702-abef-d26284fe765c

Former-commit-id: 8f874ae1e42644c122dae508225393832f3bfac9 [formerly 036b03d5cde214d1808ecce7775ec166bdf5c6ea] [formerly 319f82caf04efdff934cc2143dba5e9c6226f7a5 [formerly 346d9cf1fa8ca5e8986774275e13dda3e593f3ac]]
Former-commit-id: e8dbe2b652e0100931ede3d86fd73aadb1237dc7 [formerly 5f0b4c70fd44f1da491b300e39fe83c49c66c755]
Former-commit-id: ab27644ed76efc96d42585102d6c2e513ab40c21

+++ Source Code/Editor/Widgets/ImWindowManagerDivide.cpp
@@ -40,7 +40,7 @@ void ImwWindowManagerDivide::update(const U64 deltaTimeUS) {
 ImVec2 ImwWindowManagerDivide::GetCursorPos()
 {
     vec2<I32> cursorPos = _parent.context().app().windowManager().getCursorPosition();
-    return ImVec2(cursorPos.x, cursorPos.y);
+    return ImVec2((float)cursorPos.x, (float)cursorPos.y);
 }
 
 bool ImwWindowManagerDivide::IsLeftClickDown()
